requirements.txt
torch==2.1.0
torchvision==0.16.0
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
mlflow==2.9.0
pytorch-lightning==2.1.0